........
  r822 | weil | 2007-09-18 18:15:51 +0200 (Di, 18 Sep 2007) | 1 line
  
  cvs update
........
  r823 | weil | 2007-09-19 19:55:05 +0200 (Mi, 19 Sep 2007) | 1 line
  
  cvs update
........
  r824 | weil | 2007-09-20 20:53:32 +0200 (Do, 20 Sep 2007) | 1 line
  
  cvs update
........
  r825 | weil | 2007-09-21 20:41:47 +0200 (Fr, 21 Sep 2007) | 1 line
  
  cvs update
........
  r826 | weil | 2007-09-21 21:38:09 +0200 (Fr, 21 Sep 2007) | 1 line
  
  cvs update
........
  r827 | weil | 2007-09-22 21:15:23 +0200 (Sa, 22 Sep 2007) | 1 line
  
  cvs update
........
  r828 | weil | 2007-09-23 22:40:47 +0200 (So, 23 Sep 2007) | 1 line
  
  cvs update
........


git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/ar7-firmware/qemu/trunk@829 11fd3d49-aaf6-0310-8052-ea9566248e5e

+++ target-mips/translate.c
@@ -6417,7 +6417,7 @@ gen_intermediate_code_internal (CPUState *env, TranslationBlock *tb,
     ctx.tb = tb;
     ctx.bstate = BS_NONE;
     /* Restore delay slot state from the tb context.  */
-    ctx.hflags = tb->flags;
+    ctx.hflags = (uint32_t)tb->flags; /* FIXME: maybe use 64 bits here? */
     restore_cpu_state(env, &ctx);
 #if defined(CONFIG_USER_ONLY)
     ctx.mem_idx = 0;
